Add support for QueryDefinedTables
Added a function to support iterating through sub-queries to identify table references when sub-queries are used as table definitions. Modified if statement to identify when this was the case and call the function to recursively go through the references.

This worked on some of the example use cases I had where sub-queries were used in From statements. I have not extensively tested this functionality. I had not previously used functions to use recursion like this but used the example already there as an example.